[refactor] 카드 신청 페이지 역할 분리
- data를 관리하는 페이지와 완성된 data를 실행하는 페이지를 분리 => 역할 분리
